requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.17
alembic==1.12.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==4.6.0
requests==2.31.0
python-multipart==0.0.9
email-validator==2.1.0.post1
beautifulsoup4==4.12.3
